playground/requirements.txt
einops
gin_config
numpy
torch
wandb
matplotlib
jupyter